requirements.txt
Pillow==9.4.0
Pillow==9.5.0
replit_play==0.0.23
